requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

psycopg2

Django == 2.0.3

django_cron

djangorestframework == 3.7.7

djangorestframework_jwt == 1.11.0

django-crontab

gunicorn == 19.7.0

xhtml2pdf